Move the `ExchangeRate` component to the right place
