+++ src/modules/UI/components/ControlPanel/ControlPanel.ui.js
@@ -5,12 +5,12 @@ import { Image, Text, View } from 'react-native'
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons'
 
 import accountIcon from '../../../../assets/images/sidenav/accounts.png'
+import { ExchangeRate } from '../../../../components/common/ExchangeRate.js'
 import { SceneWrapper } from '../../../../components/common/SceneWrapper.js'
 import s from '../../../../locales/strings'
 import type { GuiDenomination } from '../../../../types/types.js'
 import { emptyGuiDenomination } from '../../../../types/types.js'
 import { getDenomFromIsoCode, getObjectDiff } from '../../../../util/utils.js'
-import ExchangeRate from '../ExchangeRate/ExchangeRate.ui.js'
 import FormattedText from '../FormattedText/FormattedText.ui.js'
 import { Button } from './Component/Button/Button.ui'
 import Main from './Component/MainConnector'
